% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_image.R
\name{add_image}
\alias{add_image}
\title{Add image to slide}
\usage{
add_image(image_url, slide_page, slides_id)
}
\arguments{
\item{image_url:}{The url to an image you would like to add to a slide}

\item{slide_page:}{the slide ID you want to add an image to}

\item{slides_id:}{The ID of slide set that contains the slide you are adding an image to}
}
\value{
factor
}
\description{
For a given slide page and slide ID, add an image to it
}
\examples{
token <- rgoogleslides::authorize()
token <- token$credentials$access_token
rgoogleslides::authorize(authorize_from_secret(token))


}
