% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize_from_secret.R
\name{authorize_from_secret}
\alias{authorize_from_secret}
\title{Authorize from access token}
\usage{
authorize_from_secret(access_token)
}
\arguments{
\item{access_token}{}
}
\value{
factor
}
\description{
Get authorization from access token
}
\examples{
token <- rgoogleslides::authorize()
token <- token$credentials$access_token
rgoogleslides::authorize(authorize_from_secret(token))

}
