% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_slide.R
\name{make_new_slide}
\alias{make_new_slide}
\title{Make a new slide}
\usage{
make_new_slide(slides_id)
}
\arguments{
\item{access_token}{}
}
\value{
factor
}
\description{
Make a new slide page with a given slides id
}
\examples{

make_new_slide(slides_id)
}
